first draft of refactor
