+++ VMEncryption/main/oscrypto/OSEncryptionState.py
@@ -29,6 +29,7 @@ from Common import *
 from CommandExecutor import *
 from BekUtil import *
 from DiskUtil import *
+from CryptMountConfigUtil import *
 from EncryptionConfig import *
 
 class OSEncryptionState(object):
@@ -46,6 +47,11 @@ class OSEncryptionState(object):
                                   patching=self.context.distro_patcher,
                                   logger=self.context.logger,
                                   encryption_environment=self.context.encryption_environment)
+
+        self.crypt_mount_config_util = CryptMountConfigUtil(hutil=self.context.hutil,
+                                                            patching=self.context.distro_patcher,
+                                                            logger=self.context.logger,
+                                                            encryption_environment=self.context.encryption_environment)
 
         self.bek_util = BekUtil(disk_util=self.disk_util,
                                 logger=self.context.logger)
